rename db2 -> db
